requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
slack-sdk==3.26.0
slack-bolt==1.18.0
pytest==7.4.3
pytest-json-report==1.5.0
pytest-html==4.1.1
allure-pytest==2.13.2
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
schedule==1.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6